pyproject.toml
[project]
name = "pythonproject12"
version = "0.1.0"
description = "Application for personal diary and planning"
authors = [
    {name = "Max",email = "z.max83@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "django (>=5.2.7,<6.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "django-filter (>=25.2,<26.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "celery (>=5.5.3,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-phonenumber-field (>=8.3.0,<9.0.0)",
    "phonenumbers (>=9.0.17,<10.0.0)",
    "drf-yasg (>=1.21.11,<2.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "django-channels (>=0.7.0,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
black = "^25.9.0"
flake8 = "^7.3.0"
isort = "^7.0.0"
mypy = "^1.18.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
coverage = "^7.11.0"

[tool.black]
line-length = 119
exclude = '.git'

[tool.mypy]
line-length = 119
disallow_untyped_defs = true
warn_return_any = true
exclude = ['venv', 'tests']

[tool.isort]
line_length = 119
